fix: Not using the header indicated in the Vary header will never use cached response.
When the server replies with a Vary but the client does not use this header, the cached response will not be used on subsequent requests. The header in the request is undefined, the header in Redis/Valkey is stored as null. Also, when no headers are used, it's never considered a match. Although the specification is unclear about the Vary header in combination with a not-used header, it also does not require the header to be present at that moment. Vary merely says something about caching subsequent requests. For example, with a Vary: Cookie the server says that if you send a cookie it should not re-use the one without cookie, but should have its own cache entry. Also, when switching from Undici-cache-redis to the builtin Undici Memory cache, it does work as in this PR.